+++ android_app/app/src/main/AndroidManifest.xml
@@ -19,29 +19,20 @@
         android:theme="@style/Theme.WallpaperWizard"
         tools:targetApi="31">
         <activity
-            android:name=".UploadActivity"
-            android:exported="false"
-            android:theme="@style/Theme.WallpaperWizard.Fullscreen">
-            <meta-data
-                android:name="android.app.lib_name"
-                android:value="" />
-        </activity>
-        <activity
             android:name=".MainActivity"
             android:exported="true"
             android:label="@string/app_name"
             android:theme="@style/Theme.WallpaperWizard.Fullscreen">
-
-            <intent-filter>
-                <action android:name="android.intent.action.MAIN" />
-                <category android:name="android.intent.category.LAUNCHER" />
-            </intent-filter>
-
             <meta-data
                 android:name="android.app.lib_name"
                 android:value="" />
-        </activity>
 
+            <intent-filter>
+                <action android:name="android.intent.action.MAIN" />
+
+                <category android:name="android.intent.category.LAUNCHER" />
+            </intent-filter>
+        </activity>
         <meta-data
             android:name="preloaded_fonts"
             android:resource="@array/preloaded_fonts" />
