+++ android_app/app/src/main/java/com/example/wallpaperwizard/NotificationProvider.kt
@@ -5,7 +5,6 @@ import android.app.NotificationManager
 import android.content.Context
 import android.os.Build
 import androidx.core.app.NotificationCompat
-import androidx.core.content.ContextCompat
 
 class NotificationProvider(private val context: Context) {
