+++ android_app/app/src/main/java/com/example/wallpaperwizard/Worker/WallpaperUploaderWorker.kt
@@ -1,16 +1,14 @@
-package com.example.wallpaperwizard
+package com.example.wallpaperwizard.Worker
 
-import android.app.Notification
 import android.app.NotificationManager
 import android.content.Context
 import android.util.Log
-import androidx.core.app.NotificationCompat
-import androidx.core.app.NotificationManagerCompat
 import androidx.core.content.ContextCompat
 import androidx.work.WorkManager
 import androidx.work.Worker
 import androidx.work.WorkerParameters
-import com.google.android.material.snackbar.Snackbar
+import com.example.wallpaperwizard.NotificationProvider
+import com.example.wallpaperwizard.WallpaperApi
 import okhttp3.MediaType.Companion.toMediaTypeOrNull
 import okhttp3.MultipartBody
 import okhttp3.RequestBody.Companion.asRequestBody
