+++ android_app/app/src/main/res/layout/activity_main.xml
@@ -1,152 +1,13 @@
 <?xml version="1.0" encoding="utf-8"?>
-<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
+<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    android:id="@+id/main_parent_view"
-    style="@style/Theme.WallpaperWizard"
+    xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    android:background="@android:color/transparent">
-
-
-    <androidx.swiperefreshlayout.widget.SwipeRefreshLayout
-        android:id="@+id/swipe_refresh_view"
+    tools:context=".MainActivity">
+    <androidx.viewpager2.widget.ViewPager2
+        android:id="@+id/pager"
         android:layout_width="match_parent"
-        android:layout_height="match_parent">
+        android:layout_height="match_parent"/>
 
-        <RelativeLayout
-            android:layout_width="match_parent"
-            android:layout_height="match_parent">
-            <LinearLayout
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:layout_centerHorizontal="true"
-                android:layout_marginTop="100dp"
-                android:gravity="center"
-                android:orientation="vertical">
-
-                <TextView
-                    android:id="@+id/date_viewer"
-                    style="@style/Widget.MaterialComponents.TextView"
-                    android:layout_width="wrap_content"
-                    android:layout_height="wrap_content"
-                    android:fontFamily="@font/arimo"
-                    android:shadowColor="@color/black"
-                    android:shadowDx="0"
-                    android:shadowDy="0"
-                    android:shadowRadius="10"
-                    android:text="Dienstag, 10. Januar"
-                    android:textColor="@color/white"
-                    android:textSize="8pt"
-                    android:textStyle="normal" />
-
-                <TextView
-                    android:id="@+id/time_viewer"
-                    style="@style/Widget.MaterialComponents.TextView"
-                    android:layout_width="wrap_content"
-                    android:layout_height="wrap_content"
-                    android:fontFamily="@font/arimo"
-                    android:shadowColor="@color/black"
-                    android:shadowDx="0"
-
-                    android:shadowDy="0"
-                    android:shadowRadius="20"
-                    android:text="12:23"
-                    android:textColor="@color/white"
-                    android:textSize="40pt"
-                    android:textStyle="bold" />
-            </LinearLayout>
-
-
-            <LinearLayout
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:layout_marginHorizontal="40dp"
-                android:layout_marginTop="270dp"
-                android:orientation="vertical">
-
-                <com.google.android.material.textview.MaterialTextView
-                    android:layout_width="wrap_content"
-                    android:layout_height="wrap_content"
-
-                    android:shadowColor="@color/black"
-                    android:shadowDx="0"
-                    android:shadowDy="0"
-                    android:shadowRadius="10"
-                    android:text="@string/sync"
-                    android:textColor="@color/white"
-                    android:textSize="10pt"
-                    android:textStyle="bold"
-                    app:fontFamily="@font/heebo" />
-
-                <com.google.android.material.textfield.TextInputEditText
-                    android:id="@+id/sync_input"
-                    android:layout_width="match_parent"
-                    android:layout_height="wrap_content"
-                    android:layout_marginTop="5dp"
-                    android:background="@drawable/input_background"
-                    android:fontFamily="@font/arimo"
-                    android:hint="@string/sync" />
-
-                <LinearLayout
-                    android:layout_width="match_parent"
-                    android:layout_height="wrap_content"
-                    android:layout_marginTop="30dp"
-                    android:orientation="vertical">
-
-                    <TextView
-                        android:layout_width="wrap_content"
-                        android:layout_height="wrap_content"
-                        android:fontFamily="@font/heebo"
-                        android:shadowColor="@color/black"
-                        android:shadowDx="0"
-                        android:shadowDy="0"
-                        android:shadowRadius="10"
-                        android:text="@string/tags"
-                        android:textColor="@color/white"
-                        android:textSize="10pt"
-                        android:textStyle="bold">
-
-
-                    </TextView>
-
-                    <com.google.android.material.chip.ChipGroup
-                        android:id="@+id/tag_group"
-                        android:layout_width="match_parent"
-                        android:layout_height="match_parent"
-                        app:singleLine="false" />
-                </LinearLayout>
-
-
-            </LinearLayout>
-
-            <com.google.android.material.floatingactionbutton.FloatingActionButton
-                android:id="@+id/applyConfig"
-                style="@style/Widget.MaterialComponents.FloatingActionButton"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:layout_alignParentEnd="true"
-                android:layout_alignParentBottom="true"
-                android:layout_marginEnd="40dp"
-                android:layout_marginBottom="48dp"
-                android:backgroundTint="@color/white"
-                android:contentDescription="@string/applyConfig"
-                android:scaleType="center"
-                android:src="@drawable/baseline_save_24" />
-
-            <com.google.android.material.floatingactionbutton.FloatingActionButton
-                android:id="@+id/addWallpaper"
-                style="@style/Widget.MaterialComponents.FloatingActionButton"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:layout_alignParentStart="true"
-                android:layout_alignParentBottom="true"
-                android:layout_marginStart="40dp"
-                android:layout_marginBottom="48dp"
-                android:backgroundTint="@color/white"
-                android:contentDescription="@string/addWallpaper"
-                android:scaleType="center"
-                android:src="@drawable/publish_48px" />
-        </RelativeLayout>
-    </androidx.swiperefreshlayout.widget.SwipeRefreshLayout>
-
-</RelativeLayout>
+</androidx.constraintlayout.widget.ConstraintLayout>
