+++ android_app/app/src/main/res/layout/fragment_upload.xml
@@ -4,20 +4,31 @@
     xmlns:tools="http://schemas.android.com/tools"
     style="@style/Theme.WallpaperWizard.Fullscreen"
     android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context=".UploadActivity">
+    android:layout_height="match_parent">
+
+
 
     <com.ortiz.touchview.TouchImageView
         android:id="@+id/wallpaper_preview"
         android:layout_width="match_parent"
         android:layout_height="match_parent" />
 
+    <View
+        android:id="@+id/scroll_overlay"
+        android:layout_width="match_parent"
+        android:layout_height="300px"
+        android:layout_above="@+id/add_tags_layout"
+        android:layout_alignParentStart="true"
+        android:layout_alignParentTop="true"
+        android:layout_alignParentEnd="true"
+        android:layout_marginBottom="0dp"/>
     <LinearLayout
+        android:id="@+id/add_tags_layout"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        android:layout_alignParentBottom="true"
         android:layout_alignParentStart="true"
         android:layout_alignParentEnd="true"
+        android:layout_alignParentBottom="true"
         android:layout_marginStart="40dp"
         android:layout_marginEnd="40dp"
         android:layout_marginBottom="500dp"
